fixed long/int warnings

+++ JapanPostBarcode.m
@@ -65,22 +65,22 @@ unsigned int japanpost_characterDescriptor(unichar character) {
 	return numberCodesOfJapan[character - '-'];
 }
 
-unsigned int japanpost_barDescriptor(unsigned int descriptor,unsigned int bar,unsigned int bars)
+NSUInteger japanpost_barDescriptor(NSUInteger descriptor,NSUInteger bar,NSUInteger bars)
 {
-	unsigned int shift = (bars - 1 - (bar / 2)) * 4; // 4ビットずつで1つのバーを表す。bar / 2 番目 //
-	unsigned int mask = 0xF << shift; // 0xfは4ビット分 //
+	NSUInteger shift = (bars - 1 - (bar / 2)) * 4; // 4ビットずつで1つのバーを表す。bar / 2 番目 //
+	NSUInteger mask = 0xF << shift; // 0xfは4ビット分 //
 
 	return (descriptor & mask) >> shift;
 }
 
-- (unsigned int)_barDescriptor:(unsigned int)index
+- (NSUInteger)_barDescriptor:(unsigned int)index
 // barcode is [self initiator][self barcode][self terminator] //
 // 2 * numberLength がほぼ定数に近い働きをしており、これを変数にすることが難しい。よって'A'~'Z'は2文字の扱いをした。 //
 {
-	unsigned int bar;
-    unsigned int hexDigit;
-    unsigned int descriptor;
-    unsigned int contentLength = [japanpostContents length];
+	NSUInteger bar;
+    NSUInteger hexDigit;
+    NSUInteger descriptor;
+    NSUInteger contentLength = [japanpostContents length];
 
     if ([[self initiator] length] > index) {
 		bar = index % [[self initiator] length];
@@ -94,7 +94,7 @@ unsigned int japanpost_barDescriptor(unsigned int descriptor,unsigned int bar,un
 			descriptor = japanpost_barDescriptor(hexDigit,bar,closeBracketLength);
 		}
 		else {
-			unsigned int digit = (index - [[self initiator] length]) / (2 * numberLength); // 何文字目か //
+			NSUInteger digit = (index - [[self initiator] length]) / (2 * numberLength); // 何文字目か //
 
 			bar = (index - [[self initiator] length]) % (2 * numberLength); // 1文字の中の何番目のバーになるか //
 			if (digit != contentLength) {
@@ -201,12 +201,12 @@ unsigned int japanpost_barDescriptor(unsigned int descriptor,unsigned int bar,un
 // 大文字化する。20文字に揃える。使用できない文字を排除する。郵便番号の整合性はチェックしない。 //
 {
 	unichar uChar;
-	unsigned int i;
+	NSUInteger i;
 	NSString *uppercaseString = [inContent uppercaseString];
 	NSMutableString *tempStr = [NSMutableString string];
 	NSMutableString *tempContents = [NSMutableString string];
 	NSString *tenStr = @"0123456789";
-	unsigned int maxLength = 20; // 郵便番号7文字+住居表示番号13文字 //
+	NSUInteger maxLength = 20; // 郵便番号7文字+住居表示番号13文字 //
 
 	for (i = 0; i < [uppercaseString length]; i++) {
 		uChar = [uppercaseString characterAtIndex:i];
@@ -242,8 +242,8 @@ unsigned int japanpost_barDescriptor(unsigned int descriptor,unsigned int bar,un
 	}
 	else {
 		if (maxLength < [tempStr length]) {
-			unsigned int validCount;
-			unsigned int tempL = [tempStr length];
+			NSUInteger validCount;
+			NSUInteger tempL = [tempStr length];
 			NSString *tempCStr = [NSString stringWithString:tempContents];
 
 			[tempStr deleteCharactersInRange:NSMakeRange(maxLength,tempL - maxLength)];
@@ -303,10 +303,10 @@ unsigned int japanpost_barDescriptor(unsigned int descriptor,unsigned int bar,un
 }
 
 - (float)barTop:(int)index {
-    return (float)(japanpost_barTop([self _barDescriptor:(unsigned int)index],(double)height) / (double)MILLIMETERPERPOINT);
+    return (float)(japanpost_barTop([self _barDescriptor:(NSUInteger)index],(double)height) / (double)MILLIMETERPERPOINT);
 }
 - (float)barBottom:(int)index {
-    return (float)(japanpost_barBottom([self _barDescriptor:(unsigned int)index],(double)height) / (double)MILLIMETERPERPOINT);    
+    return (float)(japanpost_barBottom([self _barDescriptor:(NSUInteger)index],(double)height) / (double)MILLIMETERPERPOINT);    
 }
 
 - (NSString *)_encodeChar:(char)inChar {
